backend/requirements.txt
# list of Python dependencies the project needs

django==5.1.5
# to be able to connect to postgres database
wheel==0.45.1
psycopg2==2.9.10
psycopg2-binary==2.9.10
djangorestframework==3.15.2
